create useSaveToHistory hook
